+++ backend/server.js
@@ -10,10 +10,9 @@ const app = express();
 
 // Enable CORS
 app.use(cors({
-  origin: "http://localhost:5173", // Frontend origin
-  methods: ["GET", "POST", "PUT", "DELETE"],
-  allowedHeaders: ["Content-Type", "Authorization"],
-  credentials: true, // Allow cookies if needed
+    origin: "http://localhost:5173", // Allow requests from your frontend origin
+    methods: ["GET", "POST", "PUT", "DELETE"], // Allow specific HTTP methods
+    allowedHeaders: ["Content-Type", "Authorization"], // Allow specific headers
 }));
 
 // Parse JSON bodies
@@ -59,18 +58,6 @@ app.get("/", (req, res) => {
     res.send("Server is ready");
 });
 
-app.get("/search", async (req, res) => {
-  try {
-    const { email } = req.query;
-    const query = email ? { email: email } : {}; 
-    const result = await DataModel.find(query);
-    res.status(200).json(result);
-  } catch (error) {
-    console.error("Error fetching data:", error);
-    res.status(500).json({ error: "Failed to fetch data" });
-  }
-});
-
 app.post("/save-json", async (req, res) => {
     try {
         const jsonData = req.body; // Access JSON data from the request body
@@ -80,53 +67,6 @@ app.post("/save-json", async (req, res) => {
         console.error("Error saving data:", error);
         res.status(500).json({ error: "Failed to save data" });
     }
-});
-
-app.put("/update", async (req, res) => {
-  try {
-    const { email, updates } = req.body;
-
-    if (!email || !updates) {
-      return res.status(400).json({ error: "Email and updates are required" });
-    }
-
-    const updatedData = await DataModel.findOneAndUpdate(
-      { email },
-      { $set: updates },
-      { new: true, runValidators: true }
-    );
-
-    if (!updatedData) {
-      return res.status(404).json({ message: "User not found" });
-    }
-
-    res.status(200).json({ message: "User updated successfully", updatedData });
-  } catch (error) {
-    console.error("Error updating data:", error);
-    res.status(500).json({ error: "Failed to update data" });
-  }
-});
-
-// Delete a user by email
-app.delete("/delete", async (req, res) => {
-  try {
-    const { email } = req.query;
-
-    if (!email) {
-      return res.status(400).json({ error: "Email is required" });
-    }
-
-    const deletedData = await DataModel.findOneAndDelete({ email });
-
-    if (!deletedData) {
-      return res.status(404).json({ message: "User not found" });
-    }
-
-    res.status(200).json({ message: "User deleted successfully", deletedData });
-  } catch (error) {
-    console.error("Error deleting data:", error);
-    res.status(500).json({ error: "Failed to delete data" });
-  }
 });
 
 // Start the server
