+++ frontend/src/pages/HomePage.jsx
@@ -1,0 +1,22 @@
+import React from "react";
+import { jwtDecode } from "jwt-decode";
+
+import { GoogleLogin } from "@react-oauth/google";
+
+const HomePage = () => {
+  return (
+    <GoogleLogin
+      onSuccess={async (credentialResponse) => {
+        const credentialResponseDecoded = jwtDecode(
+          credentialResponse.credential
+        );
+        console.log(credentialResponseDecoded);
+      }}
+      onError={() => {
+        console.log("Login Failed");
+      }}
+    />
+  );
+};
+
+export default HomePage;
