--- frontend/src/pages/RateProf.jsx
@@ -1,11 +1,0 @@
-
-
-function RateProf() {
-
-  return (
-    <div>
-
-    </div>
-  );
-}
-export default App;
